Refactored to more easily centralize logging, added details from env
