+++ tasks/clone.py
@@ -7,11 +7,9 @@ import shlex
 from invoke import task, call
 
 from log_utils import get_logger
+from log_utils.load_dotenv import load_dotenv
 from .common import (REPO_BASE_URL, CLONE_DIR_PATH,
                      SITE_BUILD_DIR_PATH, clean)
-
-
-LOGGER = get_logger('CLONE')
 
 
 def clone_url(owner, repository, access_token=''):  # nosec
@@ -40,6 +38,7 @@ def _clone_repo(ctx, owner, repository, branch, depth=''):
 
     Expects GITHUB_TOKEN to be in the environment.
     '''
+    LOGGER = get_logger('CLONE')
 
     LOGGER.info(f'Cloning {owner}/{repository}/{branch} to {CLONE_DIR_PATH}')
 
@@ -78,6 +77,9 @@ def push_repo_remote(ctx, owner, repository, branch,
 
     Expects GITHUB_TOKEN to be in the environment.
     '''
+    load_dotenv()
+    LOGGER = get_logger('CLONE')
+
     LOGGER.info(f'Pushing cloned repository to {owner}/{repository}/{branch}')
 
     github_token = os.environ['GITHUB_TOKEN']
